Improve UI and UX
Implement a Service ant Task from javafx.concurrent package to allow using the UI while the model is being learnt. Show the structure of a learnt model in a separate window. Allow to choose whether or not the probabilities of the class configurations are estimated during model evaluation. Add comments and clean the code.