Fix errors in evaluation metrics
The number of different classes of a class variable was retrieved from the test dataset, which may not contain all of them. They are now obtained from the entire dataset. In addition, the calculation of the precision, recall and F1 score has been fixed for cases where the denominator is 0.